requirements.txt
Django==2.1.1
django-common-helpers==0.9.2
django-cron==0.5.1
djangorestframework==3.8.2
pytz==2018.5
